Move demo out of `tinhte-api-react`
