+++ .travis.yml
@@ -2,7 +2,9 @@ sudo: false
 
 language: node_js
 node_js:
-  - '10.15.1'
+  - '14.16.0'
+addons:
+  chrome: stable
 cache:
   directories:
     - node_modules
