+++ packages/tinhte-api-demo/src/components/NewThread.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { apiHoc } from '../../../src'
+import { apiHoc } from 'tinhte-api-react'
 
 class NewThread extends React.Component {
   constructor (props) {
