+++ packages/tinhte-api-demo/src/components/VisitorThreads.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { apiHoc } from '../../../src'
+import { apiHoc } from 'tinhte-api-react'
 
 import ThreadPosts from './ThreadPosts'
